+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.0] - Unreleased
 ### Added
-- Added layout grid toggling functionality (ctrl+g)
+- Added layout grids toggling (smGrid: 'alt + ArrowLeft', lgGrid: 'alt + ArrowRight', both: 'alt + ArrowUp')
 - Added password reset functionality (<https://github.com/opencv/cvat/pull/2058>)
 - Ability to work with data on the fly (https://github.com/opencv/cvat/pull/2007)
 - Annotation in process outline color wheel (<https://github.com/opencv/cvat/pull/2084>)
