+++ cvat-ui/src/base.scss
@@ -3,7 +3,11 @@
 // SPDX-License-Identifier: MIT
 
 $grid-unit-size: 8px;
-$layout-grid-size: $grid-unit-size * 1;
+
+$layout-sm-grid-size: $grid-unit-size / 2;
+$layout-lg-grid-size: $grid-unit-size * 2;
+$layout-sm-grid-color: rgba(0, 0, 0, 0.15);
+$layout-lg-grid-color: rgba(0, 0, 0, 0.15);
 
 $header-color: #d8d8d8;
 $text-color: #303030;
@@ -27,6 +31,5 @@ $objects-bar-icons-color: #242424; // #6e6e6e
 $active-label-background-color: #d8ecff;
 $object-item-border-color: rgba(0, 0, 0, 0.7);
 $slider-color: #1890ff;
-$layout-grid-color: rgba(0, 0, 0, 0.15);
 
 $monospaced-fonts-stack: Consolas, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono, Bitstream Vera Sans Mono, Courier New, monospace;
