+++ cvat-ui/src/components/layout-grid/styles.scss
@@ -6,22 +6,44 @@
 
 .grid {
     display: grid;
-    grid-template-rows: repeat(1000, $layout-grid-size);
-    grid-template-columns: repeat(1000, $layout-grid-size);
 
     &::before,
     &::after {
         content: '';
-        position: absolute;
+        position: fixed;
         top: 0;
         left: 0;
         width: 100%;
-        height: 100%;
-        background: linear-gradient(to right, $layout-grid-color 1px, transparent 1px);
-        background-size: $layout-grid-size;
+        height: 100vw;
     }
 
-    &::after {
-        transform: rotate(90deg);
+    &.sm {
+        grid-template-rows: repeat(1000, $layout-sm-grid-size);
+        grid-template-columns: repeat(1000, $layout-sm-grid-size);
+
+        &::before,
+        &::after {
+            background: linear-gradient(to right, $layout-sm-grid-color 1px, transparent 1px);
+            background-size: $layout-sm-grid-size;
+        }
+
+        &::after {
+            transform: rotate(90deg);
+        }
+    }
+
+    &.lg {
+        grid-template-rows: repeat(1000, $layout-lg-grid-size);
+        grid-template-columns: repeat(1000, $layout-lg-grid-size);
+
+        &::before,
+        &::after {
+            background: linear-gradient(to right, $layout-lg-grid-color 1px, transparent 1px);
+            background-size: $layout-lg-grid-size;
+        }
+
+        &::after {
+            transform: rotate(90deg);
+        }
     }
 }
