+++ cvat-ui/src/styles.scss
@@ -60,4 +60,5 @@ hr {
     top: 0;
     left: 0;
     z-index: 9999;
+    pointer-events: none;
 }
